requirements.txt
Django==5.1.6
dj-database-url==2.3.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
python-dotenv==1.0.1
cloudinary==1.44.0
django-cloudinary-storage==0.3.0
Pillow==11.0.0